Update previous things

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 /* ===== Callback Style =====
    Video.find({}, (error, videos) => {
@@ -18,7 +19,9 @@ import Video from "../models/Video";
 */
 
 export const home = async (req, res) => {
-   const videos = await Video.find({}).sort({ createdAt: "desc" });
+   const videos = await Video.find({})
+      .sort({ createdAt: "desc" })
+      .populate("owner");
    return res.render("home", { pageTitle: "Home", videos });
 };
 
@@ -30,8 +33,8 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
    const { id } = req.params;
-   // const { id } = req.params === const id = req.params.id; :: ES6 style
-   const video = await Video.findById(id);
+   const video = await Video.findById(id).populate("owner");
+   // .populate("owner") === mongoose를 이용해 owner를 string이 아닌 origin data object로 제공함.
 
    if (!video) {
       return res.status(404).render("404", { pageTitle: "Video not found." });
@@ -42,10 +45,18 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
    const { id } = req.params;
+   const {
+      user: { _id },
+   } = req.session;
    const video = await Video.findById(id);
 
    if (!video) {
       return res.status(404).render("404", { pageTitle: "Video not found." });
+   }
+
+   if (String(video.owner) !== String(_id)) {
+      return res.status(403).redirect("/");
+      // status(403) === Forbidden
    }
 
    res.render("edit", { pageTitle: `Editing: ${video.title}`, video });
@@ -53,11 +64,19 @@ export const getEdit = async (req, res) => {
 
 export const postEdit = async (req, res) => {
    const { id } = req.params;
+   const {
+      user: { _id },
+   } = req.session;
    const { title, description, hashtags } = req.body;
    const video = await Video.exists({ _id: id });
 
    if (!video) {
       return res.status(404).render("404", { pageTitle: "Video not found." });
+   }
+
+   if (String(video.owner) !== String(_id)) {
+      return res.status(403).redirect("/");
+      // status(403) === Forbidden
    }
 
    await Video.findByIdAndUpdate(id, {
@@ -78,14 +97,26 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+   const {
+      user: { _id },
+   } = req.session;
+   const { path: fileUrl } = req.file;
    const { title, description, hashtags } = req.body;
 
    try {
-      await Video.create({
+      const newVideo = await Video.create({
          title,
          description,
+         fileUrl,
+         owner: _id,
          hashtags: Video.formatHashtags(hashtags),
       });
+
+      const user = await User.findById(_id);
+      user.videos.push(newVideo._id);
+      user.save();
+      // userSchema.pre("save") 때문에 hash 일어남 >> fix bug
+
       return res.redirect("/");
    } catch (error) {
       return res.status(400).render("upload", {
@@ -113,6 +144,19 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
    const { id } = req.params;
+   const {
+      user: { _id },
+   } = req.session;
+   const video = await Video.findById(id);
+
+   if (!video) {
+      return res.status(404).render("404", { pageTitle: "Video not found." });
+   }
+
+   if (String(video.owner) !== String(_id)) {
+      return res.status(403).redirect("/");
+      // status(403) === Forbidden
+   }
 
    await Video.findByIdAndDelete(id);
    // === findOneAndDelete({ _id : id })
@@ -129,7 +173,7 @@ export const search = async (req, res) => {
          title: {
             $regex: new RegExp(keyword, "i"),
          },
-      });
+      }).populate("owner");
    }
 
    return res.render("search", { pageTitle: "Search", videos });
